requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-wtf==1.2.2
wtforms==3.2.1
python-telegram-bot==22.3
pandas==2.3.1
python-dotenv==1.0.0
werkzeug==2.3.7
email-validator==2.2.0
functions-framework==3.*
requests==2.31.0
numpy==1.24.3
schedule==1.2.0